+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultLibraryDilutionService.java
@@ -153,6 +153,7 @@ public class DefaultLibraryDilutionService
   @Override
   public void update(LibraryDilution dilution) throws IOException {
     LibraryDilution managed = get(dilution.getId());
+    maybeRemoveFromBox(dilution);
     boxService.throwIfBoxPositionIsFilled(dilution);
 
     loadChildEntities(dilution);
@@ -215,6 +216,12 @@ public class DefaultLibraryDilutionService
     }
   }
 
+  private void maybeRemoveFromBox(LibraryDilution dilution) {
+    if (dilution.isDiscarded() || dilution.isDistributed()) {
+      dilution.setBoxPosition(null);
+    }
+  }
+
   /**
    * Copies modifiable fields from the source LibraryDilution into the target LibraryDilution to be persisted
    * 
@@ -225,6 +232,7 @@ public class DefaultLibraryDilutionService
   private void applyChanges(LibraryDilution target, LibraryDilution source) {
     target.setTargetedSequencing(source.getTargetedSequencing());
     target.setIdentificationBarcode(LimsUtils.nullifyStringIfBlank(source.getIdentificationBarcode()));
+    target.setDiscarded(source.isDiscarded());
     if (source.isDiscarded() || source.isDistributed()) {
       target.setVolume(0.0);
     } else {