+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/EditKitDescriptorController.java
@@ -24,9 +24,6 @@
 package uk.ac.bbsrc.tgac.miso.webapp.controller;
 
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Map;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -35,48 +32,26 @@ import org.springframework.security.acls.model.NotFoundException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.SessionAttributes;
-import org.springframework.web.bind.support.SessionStatus;
 import org.springframework.web.servlet.ModelAndView;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ArrayNode;
 
 import uk.ac.bbsrc.tgac.miso.core.data.impl.kit.KitDescriptor;
 import uk.ac.bbsrc.tgac.miso.core.data.type.KitType;
-import uk.ac.bbsrc.tgac.miso.core.data.type.PlatformType;
 import uk.ac.bbsrc.tgac.miso.dto.Dtos;
 import uk.ac.bbsrc.tgac.miso.service.KitService;
-import uk.ac.bbsrc.tgac.miso.service.security.AuthorizationManager;
 
 @Controller
 @RequestMapping("/kitdescriptor")
-@SessionAttributes("kitDescriptor")
 public class EditKitDescriptorController {
   protected static final Logger log = LoggerFactory.getLogger(EditKitDescriptorController.class);
 
   @Autowired
   private KitService kitService;
-  @Autowired
-  private AuthorizationManager authorizationManager;
-  @Autowired
-  private MenuController menuController;
-
-  @ModelAttribute("maxLengths")
-  public Map<String, Integer> maxLengths() throws IOException {
-    return kitService.getKitDescriptorColumnSizes();
-  }
-
-  @ModelAttribute("kitTypes")
-  public Collection<KitType> populateKitTypes() {
-    return Arrays.asList(KitType.values());
-  }
-
-  @ModelAttribute("platformTypes")
-  public Collection<PlatformType> populatePlatformTypes() {
-    return Arrays.asList(PlatformType.values());
-  }
 
   public void setKitService(KitService kitService) {
     this.kitService = kitService;
@@ -84,50 +59,32 @@ public class EditKitDescriptorController {
 
   @GetMapping(value = "/new")
   public ModelAndView setupForm(ModelMap model) throws IOException {
-    model.addAttribute("kitDescriptor", null);
-    return setupForm(KitDescriptor.UNSAVED_ID, model);
+    model.addAttribute("title", "New Kit Descriptor");
+    return setupForm(new KitDescriptor(), model);
   }
 
   @GetMapping(value = "/{kitDescriptorId}")
   public ModelAndView setupForm(@PathVariable Long kitDescriptorId, ModelMap model) throws IOException {
-    try {
-      KitDescriptor kitDescriptor = null;
-      if (kitDescriptorId == KitDescriptor.UNSAVED_ID) {
-        kitDescriptor = new KitDescriptor();
-        model.addAttribute("title", "New Kit Descriptor");
-      } else {
-        kitDescriptor = kitService.getKitDescriptorById(kitDescriptorId);
-        if (kitDescriptor == null) throw new NotFoundException("No kit found for ID " + kitDescriptorId.toString());
-        model.put("title", "Kit Descriptor " + kitDescriptor.getId());
-      }
-
-      model.put("formObj", kitDescriptor);
-      model.put("kitDescriptor", kitDescriptor);
-      model.put("associatedTargetedSequencings", Dtos.asTargetedSequencingDtos(kitDescriptor.getTargetedSequencing()));
-      model.put("isUserAdmin", authorizationManager.getCurrentUser().isAdmin());
-      return new ModelAndView("/WEB-INF/pages/editKitDescriptor.jsp", model);
-    } catch (IOException ex) {
-      if (log.isDebugEnabled()) {
-        log.debug("Failed to show Kit Descriptor", ex);
-      }
-      throw ex;
+    KitDescriptor kitDescriptor = kitService.getKitDescriptorById(kitDescriptorId);
+    if (kitDescriptor == null) {
+      throw new NotFoundException("No kit found for ID " + kitDescriptorId.toString());
     }
+    model.put("title", "Kit Descriptor " + kitDescriptor.getId());
+    return setupForm(kitDescriptor, model);
   }
 
-  @PostMapping
-  public ModelAndView processSubmit(@ModelAttribute("kitDescriptor") KitDescriptor kitDescriptor, ModelMap model, SessionStatus session)
-      throws IOException {
-    try {
-      kitService.saveKitDescriptor(kitDescriptor);
-      menuController.refreshConstants();
-      session.setComplete();
-      model.clear();
-      return new ModelAndView("redirect:/miso/kitdescriptor/" + kitDescriptor.getId(), model);
-    } catch (IOException ex) {
-      if (log.isDebugEnabled()) {
-        log.debug("Failed to save Kit Descriptor", ex);
-      }
-      throw ex;
+  public ModelAndView setupForm(KitDescriptor kitDescriptor, ModelMap model) throws JsonProcessingException {
+    model.put("kitDescriptor", kitDescriptor);
+    model.put("associatedTargetedSequencings", Dtos.asTargetedSequencingDtos(kitDescriptor.getTargetedSequencing()));
+
+    ObjectMapper mapper = new ObjectMapper();
+    model.put("kitDescriptorDto", mapper.writeValueAsString(Dtos.asDto(kitDescriptor)));
+    ArrayNode kitTypes = mapper.createArrayNode();
+    for (String item : KitType.getKeys()) {
+      kitTypes.add(item);
     }
+    model.put("kitTypes", kitTypes);
+
+    return new ModelAndView("/WEB-INF/pages/editKitDescriptor.jsp", model);
   }
 }