+++ RELEASE_NOTES.md
@@ -2,11 +2,15 @@
 
 Changes:
 
+  * Added discarded and box location fields to Edit Dilution page
   * Changed Edit Run page to show validation messages instead of error page when saving fails
   * Changed Edit Service Record page to show validation messages instead of error page when saving
     fails
   * Changed Edit Kit Descriptor page to show validation messages instead of error page when saving
     fails
+  * Changed the Edit Project page to show validation messages instead of error page when saving
+    fails
+  * Changed Edit Workset page to show validation messages instead of error page when saving fails
   * Made dilutions and pools distributable
   * Improved "Location" display in list tables
   * Fixed Pinery-MISO failing to reverse compliment indices when appropriate if the container model